+++ frontend/src/libs/packages/store/store.package.ts
@@ -7,7 +7,7 @@ import { configureStore } from '@reduxjs/toolkit';
 
 import { AppEnvironment } from '#libs/enums/enums.js';
 import { type Config } from '#libs/packages/config/config.js';
-import { errorHandlingMiddleware } from '#middlewares/errorHandling/error-handling-middleware.js';
+import { errorHandlingMiddleware } from '#middlewares/error-handling/error-handling-middleware.js';
 import { authApi } from '#packages/auth/auth.js';
 import { userApi } from '#packages/users/users.js';
 import { reducer as authReducer } from '#slices/auth/auth.js';
